Update Prisma schema with new fields and enums
Added 'category' and 'logChannelId' fields to SupportGuilds and introduced the GuildCategory enum. Updated StaffUsers with a 'departments' JSON field and added 'Trial_Administrator' to StaffRoles. Modified Report and Appeal models to include 'transcriptId' and added 'channelId' to Appeal with a unique constraint. Cleaned up and reorganized some model fields for clarity.